+++ smartcontract/programs/doublezero-telemetry/tests/initialize_device_latency_samples_tests.rs
@@ -162,6 +162,209 @@ async fn test_initialize_device_latency_samples_already_with_lamports() {
 }
 
 #[tokio::test]
+async fn test_initialize_device_latency_samples_success_suspended_origin_device() {
+    let mut ledger = LedgerHelper::new().await.unwrap();
+
+    let payer_pubkey = ledger
+        .context
+        .lock()
+        .unwrap()
+        .payer
+        .insecure_clone()
+        .pubkey();
+    let contributor_pk = ledger
+        .serviceability
+        .create_contributor("CONTRIB".to_string(), payer_pubkey)
+        .await
+        .unwrap();
+
+    // Seed ledger with two linked devices, and a funded origin device agent.
+    let (origin_device_agent, origin_device_pk, target_device_pk, link_pk) = ledger
+        .seed_with_two_linked_devices(contributor_pk)
+        .await
+        .unwrap();
+
+    // Drain the origin device.
+    ledger
+        .serviceability
+        .softdrained_device(contributor_pk, origin_device_pk)
+        .await
+        .unwrap();
+
+    // Wait for a new blockhash before moving on.
+    ledger.wait_for_new_blockhash().await.unwrap();
+
+    // Check that the origin device is suspended.
+    let device = ledger
+        .serviceability
+        .get_device(origin_device_pk)
+        .await
+        .unwrap();
+    assert_eq!(device.status, DeviceStatus::SoftDrained);
+
+    // Execute initialize latency samples transaction.
+    let latency_samples_pda = ledger
+        .telemetry
+        .initialize_device_latency_samples(
+            &origin_device_agent,
+            origin_device_pk,
+            target_device_pk,
+            link_pk,
+            1u64,
+            5_000_000,
+        )
+        .await
+        .unwrap();
+
+    // Verify account creation and data.
+    let account = ledger
+        .get_account(latency_samples_pda)
+        .await
+        .unwrap()
+        .unwrap();
+    assert_eq!(account.owner, ledger.telemetry.program_id);
+    assert_eq!(account.data.len(), DEVICE_LATENCY_SAMPLES_HEADER_SIZE);
+    assert_eq!(
+        account.lamports,
+        EXPECTED_LAMPORTS_USED_FOR_ACCOUNT_CREATION
+    );
+}
+
+#[tokio::test]
+async fn test_initialize_device_latency_samples_success_suspended_target_device() {
+    let mut ledger = LedgerHelper::new().await.unwrap();
+
+    let payer_pubkey = ledger
+        .context
+        .lock()
+        .unwrap()
+        .payer
+        .insecure_clone()
+        .pubkey();
+    let contributor_pk = ledger
+        .serviceability
+        .create_contributor("CONTRIB".to_string(), payer_pubkey)
+        .await
+        .unwrap();
+
+    // Seed ledger with two linked devices, and a funded origin device agent.
+    let (origin_device_agent, origin_device_pk, target_device_pk, link_pk) = ledger
+        .seed_with_two_linked_devices(contributor_pk)
+        .await
+        .unwrap();
+
+    // Suspend the target device.
+    ledger
+        .serviceability
+        .softdrained_device(contributor_pk, target_device_pk)
+        .await
+        .unwrap();
+
+    // Wait for a new blockhash before moving on.
+    ledger.wait_for_new_blockhash().await.unwrap();
+
+    // Check that the target device is suspended.
+    let device = ledger
+        .serviceability
+        .get_device(target_device_pk)
+        .await
+        .unwrap();
+    assert_eq!(device.status, DeviceStatus::SoftDrained);
+
+    // Execute initialize latency samples transaction.
+    let latency_samples_pda = ledger
+        .telemetry
+        .initialize_device_latency_samples(
+            &origin_device_agent,
+            origin_device_pk,
+            target_device_pk,
+            link_pk,
+            1u64,
+            5_000_000,
+        )
+        .await
+        .unwrap();
+
+    // Verify account creation and data.
+    let account = ledger
+        .get_account(latency_samples_pda)
+        .await
+        .unwrap()
+        .unwrap();
+    assert_eq!(account.owner, ledger.telemetry.program_id);
+    assert_eq!(account.data.len(), DEVICE_LATENCY_SAMPLES_HEADER_SIZE);
+    assert_eq!(
+        account.lamports,
+        EXPECTED_LAMPORTS_USED_FOR_ACCOUNT_CREATION
+    );
+}
+
+#[tokio::test]
+async fn test_initialize_device_latency_samples_success_suspended_link() {
+    let mut ledger = LedgerHelper::new().await.unwrap();
+
+    let payer_pubkey = ledger
+        .context
+        .lock()
+        .unwrap()
+        .payer
+        .insecure_clone()
+        .pubkey();
+    let contributor_pk = ledger
+        .serviceability
+        .create_contributor("CONTRIB".to_string(), payer_pubkey)
+        .await
+        .unwrap();
+
+    // Seed ledger with two linked devices, and a funded origin device agent.
+    let (origin_device_agent, origin_device_pk, target_device_pk, link_pk) = ledger
+        .seed_with_two_linked_devices(contributor_pk)
+        .await
+        .unwrap();
+
+    // Suspend the link.
+    ledger
+        .serviceability
+        .soft_drain_link(contributor_pk, link_pk)
+        .await
+        .unwrap();
+
+    // Wait for a new blockhash before moving on.
+    ledger.wait_for_new_blockhash().await.unwrap();
+
+    // Check that the link is suspended.
+    let link = ledger.serviceability.get_link(link_pk).await.unwrap();
+    assert_eq!(link.status, LinkStatus::SoftDrained);
+
+    // Execute initialize latency samples transaction.
+    let latency_samples_pda = ledger
+        .telemetry
+        .initialize_device_latency_samples(
+            &origin_device_agent,
+            origin_device_pk,
+            target_device_pk,
+            link_pk,
+            1u64,
+            5_000_000,
+        )
+        .await
+        .unwrap();
+
+    // Verify account creation and data.
+    let account = ledger
+        .get_account(latency_samples_pda)
+        .await
+        .unwrap()
+        .unwrap();
+    assert_eq!(account.owner, ledger.telemetry.program_id);
+    assert_eq!(account.data.len(), DEVICE_LATENCY_SAMPLES_HEADER_SIZE);
+    assert_eq!(
+        account.lamports,
+        EXPECTED_LAMPORTS_USED_FOR_ACCOUNT_CREATION
+    );
+}
+
+#[tokio::test]
 async fn test_initialize_device_latency_samples_fail_unauthorized_agent() {
     let mut ledger = LedgerHelper::new().await.unwrap();
 
@@ -663,7 +866,7 @@ async fn test_initialize_device_latency_samples_fail_origin_device_not_activated
         )
         .await;
 
-    assert_telemetry_error(result, TelemetryError::DeviceNotActiveOrSuspended);
+    assert_telemetry_error(result, TelemetryError::DeviceNotActivated);
 }
 
 #[tokio::test]
@@ -799,7 +1002,7 @@ async fn test_initialize_device_latency_samples_fail_target_device_not_activated
         )
         .await;
 
-    assert_telemetry_error(result, TelemetryError::DeviceNotActiveOrSuspended);
+    assert_telemetry_error(result, TelemetryError::DeviceNotActivated);
 }
 
 #[tokio::test]
@@ -931,7 +1134,7 @@ async fn test_initialize_device_latency_samples_fail_link_not_activated() {
         )
         .await;
 
-    assert_telemetry_error(result, TelemetryError::LinkNotActiveOrSuspended);
+    assert_telemetry_error(result, TelemetryError::LinkNotActivated);
 }
 
 #[tokio::test]